pyproject.toml
[tool.poetry]
name = "storyBot"
version = "0.1.0"
description = ""
authors = ["Roman Kliotzkin <roman.kliotzkin@deliverect.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.2"
pymongo = "^4.5.0"
uvicorn = "^0.23.2"
python-dotenv = "^1.0.0"
python3-saml = "^1.16.0"
itsdangerous = "^2.1.2"
python-multipart = "^0.0.6"
prometheus-fastapi-instrumentator = "^6.1.0"
openai= "^1.23.3"
google-generativeai="0.5.2"
markdown = "^3.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

